wasm-syntax, wasm: must import DOMAINS to be able to parse domains.k
This issue was made apparent by the switch to RV-K, as it does not auto-import modules for us.
It seems rather silly that we would force the modules in domains.k to be parsed by `--syntax-module`, instead of themselves.
Because of this, several of the modules in domains.k were not parsing correctly in `WASM-SYNTAX`.