Bugfix: implement correct branching and label semantics (#84)
* Bugfix: implement correct branching and label semantics

Previosuly, we always grabbed the continuation whenever we encountered a
label. However, the correct semantics is that you grab the continuation
when *branching* to a label. When it is encounterd without a branch,
it's discarded.

The correct way to write an infinite loop in wasm is

```
loop
  (nop) ;; or whatever
  (br 0)
end
```

but before this fix the following was also an infinite loop

```
loop
  (nop) ;; or whatever
end
```

* Use side conditions in `br_if`

* Update `simple_arithmetic_spec.k` to loop correctly