Bugfix: `elem` should initialize with function addresses, not ids
Previusly we were storing identifiers, but two modules may use the same
identifier for the same function, and the lookup would thus break.

For example
```
  (table (import "Mt" "tab") 5 funcref)
  (elem (i32.const 1) $i $h)
```
This should put the functions `$i` and `$h` in the *current* module in the
table. Note, though, the the module where the table is stored may call into the
table. But it doesn't have access to the identifiers `$i` and `$h` (they are
module local).