Bugfix: `elem` should initialize with function addresses, not ids (#127)
* Bugfix: `elem` should initialize with function addresses, not ids

Previusly we were storing identifiers, but two modules may use the same
identifier for the same function, and the lookup would thus break.

For example
```
(module $0
  (table (export "tab") 5 funcref)
  (type (func (result i32)))
  (func $f)
  (func (export "call") (type 0) (call_indirect (type 0) (i32.const 0)))
)

(register "Mt" $0)

(module $1
  (table (import "Mt" "tab") 5 funcref)
  (elem (i32.const 0) $f)
  (func $f (result i32) (i32.const 42))
)

(invoke $0 "call")
```

This should put the function `$f` in from *current* module in the
table. Note, though, the module where the table is stored may call into the
table. If the table stored identifiers, the `call_indirect` instruction would find the id `$f` and call the function of that name defined in $0, which is incorrect.

* Update tests

* Formatting

* Bugfix: look up global address, not local address

* Bugfix: call_indirect fix seems to have been lost in merge (?)

* Formatting

* a tiny fix