"Correct" filename for kompiling to Haskell
Kompiling to haskell no longer produces a `kore.txt` file, but rather a
`compiled.txt` file. Due to the nature of Makefiles, this issue meant
that make constantly found the `kore.txt` file lacking, and thus always
rebuilt the haskell backend regardless of if it's necessary or not.
That's a lot of wasted time every time `make build` is called.